See the details
Corrected errors in deletedatabaseelement.
Added a way to delete everything in the database linked to the current
instance of the module on deletion of this instance.
Changed the way failure works, now it no longer updates the database.
Clicking on the name of a building also redirect to the list of its
rooms.
Commentaries now have a big 'edit' button.